requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
celery==5.4.0
redis==5.1.1
django-celery-beat==2.7.0
psycopg2-binary==2.9.9
cryptography==43.0.3
django-encrypted-model-fields==0.6.5
requests==2.32.3
python-decouple==3.8
pillow==11.0.0
boto3==1.35.72
gunicorn==23.0.0
django-filter==24.3
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
